WIP - support dev w/ local Snapyr Android SDK
Adds a check for an ext value, `SnapyrRnSdk_omitSnapyrAndroidSDK`.
If this is set to true on the consumer root project, this RN SDK
will not include the Snapyr Android SDK as a dependency.

It's up to the consuming app to add Android SDK as a dependency
itself if this flag is set - if not, building will fail.