Disable `createReactContextInBackground()` call...
... in broadcast receivers. This can cause RN code to run in the
background if a push notification arrives when app was not already
running. If the app was not prepared for that it may cause
unexpected behavior.

We're not using this yet (but it would allow us to support JS code
running in reaction to push notifications). TODO: make optional/
configurable, then re-enable based on config